feat(core): ajoute les exports du coeur
