+++ src/entities/core/types/amplifyBaseTypes.ts
@@ -1,6 +1,18 @@
-// src/types/modelForm.ts
+// src/entities/core/types/amplifyBaseTypes.ts
+
 import type { Schema } from "@/amplify/data/resource";
-import type { CreateOmit } from "@myTypes/amplifyBaseTypes";
+
+/**
+ * Utilitaire générique pour tous les types basés sur un modèle Schema
+ */
+export type BaseModel<K extends keyof Schema> = Schema[K]["type"];
+
+export type CreateOmit<K extends keyof Schema> = Omit<
+    BaseModel<K>,
+    "id" | "createdAt" | "updatedAt"
+>;
+
+export type UpdateInput<K extends keyof Schema> = Partial<CreateOmit<K>>;
 
 /**
  * Type générique représentant la structure d'un formulaire basé sur un modèle,
@@ -19,17 +31,3 @@ export type ModelForm<
     CTMap extends Record<string, unknown> = Record<string, never>,
     CT extends keyof CTMap = never,
 > = Omit<CreateOmit<K>, O | R | CT> & { [P in CT]: CTMap[P] } & { [P in R as `${P}Ids`]: string[] };
-
-/**
- * Génère un couple { initialForm, toForm } pour un modèle spécifique.
- * Cela permet de centraliser la logique de transformation des données.
- */
-export function createModelForm<M, F, A extends unknown[] = unknown[]>(
-    initialForm: F,
-    toForm: (model: M, ...args: A) => F
-) {
-    return {
-        initialForm,
-        toForm,
-    } as const;
-}
