+++ src/entities/models/section/types.ts
@@ -1,6 +1,5 @@
-import type { BaseModel, CreateOmit, UpdateInput, ModelForm } from "@myTypes/amplifyBaseTypes";
+import type { BaseModel, CreateOmit, UpdateInput, ModelForm } from "@entities/core";
 import { type SeoTypeOmit } from "@src/entities";
-
 
 export type SectionTypes = BaseModel<"Section">;
 export type SectionTypesOmit = CreateOmit<"Section">;
