+++ src/entities/models/userName/form.ts
@@ -1,4 +1,4 @@
-import { createModelForm } from "@utils/createModelForm";
+import { createModelForm } from "@entities/core";
 import type { UserNameType, UserNameFormType } from "./types";
 
 export const { initialForm: initialUserNameForm, toForm: toUserNameForm } = createModelForm<
