+++ src/entities/models/userProfile/form.ts
@@ -1,4 +1,4 @@
-import { createModelForm } from "@utils/createModelForm";
+import { createModelForm } from "@entities/core";
 import type { UserProfileType, UserProfileFormType } from "./types";
 
 export const { initialForm: initialUserProfileForm, toForm: toUserProfileForm } = createModelForm<
