+++ src/entities/relations/postTag/types.ts
@@ -1,4 +1,4 @@
-import type { BaseModel, CreateOmit, UpdateInput } from "@myTypes/amplifyBaseTypes";
+import type { BaseModel, CreateOmit, UpdateInput } from "@entities/core";
 
 export type PostTagType = BaseModel<"PostTag">;
 export type PostTagTypeOmit = CreateOmit<"PostTag">;
