+++ src/entities/relations/sectionPost/types.ts
@@ -1,4 +1,4 @@
-import type { BaseModel, CreateOmit, UpdateInput } from "@myTypes/amplifyBaseTypes";
+import type { BaseModel, CreateOmit, UpdateInput } from "@entities/core";
 
 export type SectionPostType = BaseModel<"SectionPost">;
 export type SectionPostTypeOmit = CreateOmit<"SectionPost">;
