+++ tsconfig.json
@@ -22,7 +22,6 @@
             "@components/*": ["src/components/*"],
             "@hooks/*": ["src/hooks/*"],
             "@context/*": ["src/context/*"],
-            "@utils/*": ["src/utils/*"],
             "@services/*": ["src/services/*"],
             "@myTypes/*": ["src/types/*"],
             "@entities/*": ["src/entities/*"],
